ocaml_import: add 'all' attributee
* to (hermetically) import an archive we need to also import the
submodule cmx and cmi files, so we can add them to action inputs.
* previously we were cheating by passing "special" -I args with
initial '+' char, that the compiler could interpret relative to the
stdlib; this allowed it to bypass Bazel's caching and access control.
* now we've switched to having one repo per opam pkg, so we can't
cheat. or at least it would be harder. in any case, to import an
archive, we put the .cmxa and .a files in the 'archive' attr, and glob
all *.cmx, *.cmi files in the 'all' attr. The archives will end up on
the cmd line for linking, the 'all' files will go in the action inputs.